Fix lag when user has no stacks
Eliminates an effect loop related to `stackId` query param by
eliminating that parameter. It was only used for testing, and had no
value.
