+++ apps/desktop/cypress/e2e/unifiedDiffView.cy.ts
@@ -17,8 +17,6 @@ describe('Unified Diff View', () => {
 		);
 
 		cy.visit('/');
-
-		cy.url({ timeout: 3000 }).should('include', `/workspace?stackId=${mockBackend.stackId}`);
 	});
 
 	afterEach(() => {
