+++ apps/desktop/cypress/e2e/upstreamIntegration.cy.ts
@@ -17,7 +17,7 @@ describe('Upstream Integration', () => {
 
 		cy.visit('/');
 
-		cy.url({ timeout: 3000 }).should('include', `/workspace?stackId=${mockBackend.stackId}`);
+		cy.url({ timeout: 3000 }).should('include', `/${PROJECT_ID}/workspace`);
 	});
 
 	afterEach(() => {
