+++ apps/desktop/src/routes/[projectId]/workspace/+page.svelte
@@ -1,45 +1,60 @@
 <script lang="ts">
 	import { goto } from '$app/navigation';
 	import { page } from '$app/state';
-	import ReduxResult from '$components/ReduxResult.svelte';
-	import StackDraft from '$components/v3/StackDraft.svelte';
-	import noBranchesSvg from '$lib/assets/empty-state/no-branches.svg?raw';
-	import { stackPath } from '$lib/routes/routes.svelte';
+	import WorkspaceView from '$components/v3/WorkspaceView.svelte';
+	import { SettingsService } from '$lib/config/appSettingsV2';
+	import { ModeService } from '$lib/mode/modeService';
 	import { StackService } from '$lib/stacks/stackService.svelte';
 	import { UiState } from '$lib/state/uiState.svelte';
-	import { inject } from '@gitbutler/shared/context';
-	import EmptyStatePlaceholder from '@gitbutler/ui/EmptyStatePlaceholder.svelte';
+	import { getContext, inject } from '@gitbutler/shared/context';
+	import { QueryStatus } from '@reduxjs/toolkit/query';
+	import type { PageData } from './$types';
 
-	const projectId = page.params.projectId!;
+	const settingsService = getContext(SettingsService);
+	const modeService = getContext(ModeService);
+	const settingsStore = settingsService.appSettings;
+	const mode = modeService.mode;
 
+	const { data }: { data: PageData } = $props();
+
+	const projectId = $derived(page.params.projectId!);
 	const [uiState, stackService] = inject(UiState, StackService);
+	const projectState = $derived(uiState.project(projectId));
+	const stackId = $derived(projectState.stackId.current);
 
-	const drawer = $derived(uiState.project(projectId).drawerPage);
+	const firstStackResult = $derived(stackService.stackAt(projectId, 0));
+	const firstStack = $derived(firstStackResult.current.data);
+	const stackResult = $derived(stackId ? stackService.stackById(projectId, stackId) : undefined);
 
-	const stacksResult = $derived(stackService.stacks(projectId));
-	const isCommitting = $derived(drawer.current === 'new-commit');
+	$effect(() => {
+		if (stackId === undefined && firstStack) {
+			projectState.stackId.set(firstStack.id);
+		} else if (
+			stackId &&
+			stackResult?.current.status === QueryStatus.fulfilled &&
+			stackResult.current.data === undefined
+		) {
+			projectState.stackId.set(undefined);
+		}
+	});
+
+	// Redirect to board if we have switched away from V3 feature.
+	$effect(() => {
+		if ($settingsStore && !$settingsStore.featureFlags.v3) {
+			goto(`/${data.projectId}/board`);
+		}
+	});
+
+	function gotoEdit() {
+		goto(`/${projectId}/edit`);
+	}
+
+	$effect(() => {
+		if ($mode?.type === 'Edit') {
+			// That was causing an incorrect linting error when project.id was accessed inside the reactive block
+			gotoEdit();
+		}
+	});
 </script>
 
-<ReduxResult {projectId} result={stacksResult.current}>
-	{#snippet children(stacks, env)}
-		{#if stacks.length > 0}
-			{goto(stackPath(env.projectId, stacks[0]!.id))}
-		{:else if isCommitting}
-			<StackDraft {projectId} />
-		{:else}
-			<EmptyStatePlaceholder
-				image={noBranchesSvg}
-				background="none"
-				bottomMargin={40}
-				topBottomPadding={0}
-			>
-				{#snippet title()}
-					You have no branches
-				{/snippet}
-				{#snippet caption()}
-					Create a new branch for<br />a feature, fix, or idea!
-				{/snippet}
-			</EmptyStatePlaceholder>
-		{/if}
-	{/snippet}
-</ReduxResult>
+<WorkspaceView {projectId} {stackId} />
